Unified translate udf to work with both single coding and arrays of codings.
