+++ fhir-server/src/main/java/au/csiro/pathling/fhirpath/TerminologyUtils.java
@@ -20,7 +20,13 @@ package au.csiro.pathling.fhirpath;
 import au.csiro.pathling.fhirpath.element.ElementPath;
 import au.csiro.pathling.fhirpath.literal.CodingLiteralPath;
 import javax.annotation.Nonnull;
+import au.csiro.pathling.utilities.Preconditions;
+import org.apache.spark.sql.Column;
+import org.apache.spark.sql.Dataset;
+import org.apache.spark.sql.Row;
 import org.hl7.fhir.r4.model.Enumerations.FHIRDefinedType;
+
+import static au.csiro.pathling.sql.Terminology.member_of;
 
 /**
  * Terminology helper functions
@@ -54,5 +60,15 @@ public interface TerminologyUtils {
     } else {
       return false;
     }
+  }
+
+  @Nonnull
+  static Column getCodingColumn(@Nonnull final FhirPath fhirPath) {
+    Preconditions.check(isCodingOrCodeableConcept(fhirPath),
+        "Coding or CodeableConcept path expected");
+    final Column conceptColumn = fhirPath.getValueColumn();
+    return isCodeableConcept(fhirPath)
+           ? conceptColumn.getField("coding")
+           : conceptColumn;
   }
 }
