+++ terminology/src/main/java/au/csiro/pathling/sql/udf/TerminologyUdfRegistrar.java
@@ -1,11 +1,10 @@
 package au.csiro.pathling.sql.udf;
 
 import au.csiro.pathling.terminology.TerminologyServiceFactory;
-import org.apache.spark.sql.SparkSession;
-import javax.annotation.Nonnull;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import javax.annotation.Nonnull;
+import org.apache.spark.sql.SparkSession;
 
 public class TerminologyUdfRegistrar extends SqlFunctionRegistrar {
 
@@ -13,9 +12,8 @@ public class TerminologyUdfRegistrar extends SqlFunctionRegistrar {
     super(Collections.emptyList(),
         List.of(new ValidateCoding(tsf)),
         Collections.emptyList(),
-        Arrays.asList(new TranslateCoding(tsf), new TranslateCodingArray(tsf)));
+        List.of(new TranslateCoding(tsf)));
   }
-
 
   public static void registerUdfs(@Nonnull final SparkSession spark,
       @Nonnull final TerminologyServiceFactory terminologyServiceFactory) {
