--- terminology/src/main/java/au/csiro/pathling/sql/udf/TranslateCodingArray.java
@@ -1,37 +1,0 @@
-package au.csiro.pathling.sql.udf;
-
-import static au.csiro.pathling.sql.udf.TerminologyUdfHelpers.decodeMany;
-import static au.csiro.pathling.sql.udf.TerminologyUdfHelpers.encodeMany;
-
-import au.csiro.pathling.terminology.TerminologyServiceFactory;
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.spark.sql.Row;
-import scala.collection.mutable.WrappedArray;
-
-@Slf4j
-public class TranslateCodingArray extends TranslateCodingBase implements
-    SqlFunction4<WrappedArray<Row>, String, Boolean, String, Row[]> {
-
-  private static final long serialVersionUID = 7605853352299165569L;
-
-  public static final String FUNCTION_NAME = "translate_coding_array";
-
-  public TranslateCodingArray(@Nonnull final TerminologyServiceFactory terminologyServiceFactory) {
-    super(terminologyServiceFactory);
-  }
-
-  @Override
-  public String getName() {
-    return FUNCTION_NAME;
-  }
-
-  @Nullable
-  @Override
-  public Row[] call(@Nullable final WrappedArray<Row> codingsArrayRow,
-      @Nullable final String conceptMapUri, @Nullable Boolean reverse,
-      @Nullable final String equivalences) throws Exception {
-    return encodeMany(doCall(decodeMany(codingsArrayRow), conceptMapUri, reverse, equivalences));
-  }
-}
